+++ plutarch-ledger-api/src/Plutarch/LedgerApi/V1/Credential.hs
@@ -28,8 +28,6 @@ data PCredential (s :: S)
     , -- | @since 2.0.0
       PEq
     , -- | @since 2.0.0
-      PPartialOrd
-    , -- | @since 2.0.0
       POrd
     , -- | @since 2.0.0
       PShow
@@ -74,8 +72,6 @@ data PStakingCredential (s :: S)
       PIsData
     , -- | @since 2.0.0
       PEq
-    , -- | @since 2.0.0
-      PPartialOrd
     , -- | @since 2.0.0
       POrd
     , -- | @since 2.0.0